requirements.txt
Flask==3.1.3
gunicorn==25.1.0
langdetect==1.0.9
nltk==3.9.2
numpy==2.4.2
pandas==3.0.1
requests==2.32.5
scikit-learn==1.8.0
vaderSentiment==3.3.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.47
joblib
sentence-transformers
huggingface-hub
werkzeug
Flask-Limiter

python-dotenv
